+++ src/MySQLReplication/Repository/RepositoryInterface.php
@@ -3,17 +3,17 @@
 namespace MySQLReplication\Repository;
 
 /**
- * Class MySQLRepository
+ * Interface RepositoryInterface
  * @package MySQLReplication\Repository
  */
 interface RepositoryInterface
 {
     /**
-     * @param string $schema
+     * @param string $database
      * @param string $table
      * @return array
      */
-    public function getFields($schema, $table);
+    public function getFields($database, $table);
 
     /**
      * @return mixed